+++ code/core/src/manager/components/sidebar/TagsFilter.stories.tsx
@@ -1,139 +1,14 @@
-import React from 'react';
-
-import { Channel } from 'storybook/internal/channels';
-import type {
-  API_Provider,
-  DecoratorFunction,
-  DocsIndexEntry,
-  StoryIndexEntry,
-} from 'storybook/internal/types';
+import type { DocsIndexEntry, StoryIndexEntry } from 'storybook/internal/types';
 
 import { global } from '@storybook/global';
 
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
-import { deepMerge } from '@vitest/utils';
-import type { API, State } from 'storybook/manager-api';
-import { expect, fn, screen, waitFor } from 'storybook/test';
+import type { API } from 'storybook/manager-api';
+import { expect, screen, waitFor } from 'storybook/test';
 
-import type { ModuleArgs, ModuleFn } from '../../../manager-api/lib/types';
-import { init as initLayout } from '../../../manager-api/modules/layout';
-import { createTestingStore } from '../../../manager-api/store';
 import { TagsFilter } from './TagsFilter';
-
-/** Mock API wrapper that forces component updates when store state changes. */
-export class MockAPIWrapper<SubAPI, SubState> extends React.Component<{
-  children: React.ReactNode;
-  args: Record<string, unknown>;
-  initFn: ModuleFn<SubAPI, SubState>;
-  initOptions?: Partial<ModuleArgs>;
-  initialStoryState?: Partial<State>;
-}> {
-  api: ReturnType<typeof initLayout>['api'];
-  store: ReturnType<typeof createTestingStore>;
-  channel: API_Provider<API>['channel'];
-  mounted: boolean;
-
-  constructor(props: {
-    children: React.ReactNode;
-    args: Record<string, unknown>;
-    initFn: ModuleFn<SubAPI, SubState>;
-    initOptions?: Partial<ModuleArgs>;
-    initialStoryState?: Partial<State>;
-  }) {
-    super(props);
-
-    // Set up store.
-    this.mounted = false;
-    this.store = createTestingStore({} as State, (newState) => {
-      if (this.mounted) {
-        this.setState(newState);
-      }
-    });
-
-    // Mock channel and provider.
-    this.channel = new Channel({}) satisfies API_Provider<API>['channel'];
-    const provider: API_Provider<API> = {
-      getConfig: () => ({}),
-      handleAPI: () => {},
-      channel: this.channel,
-    };
-
-    // Mock other submodules we depend on.
-    const fullAPI = {
-      experimental_setFilter: fn().mockName('API::experimental_setFilter'),
-    } as unknown as API;
-
-    const { api, init, state } = props.initFn({
-      fullAPI,
-      store: this.store,
-      provider,
-      location: { search: '' },
-      navigate: () => {},
-      path: '',
-      docsOptions: {},
-      state: {} as State,
-      ...(props.initOptions ?? {}),
-    });
-
-    // Apply module and initial story states.
-    if (props.initialStoryState) {
-      this.store.setState(deepMerge<State>(state as State, props.initialStoryState));
-    } else {
-      this.store.setState(state as State);
-    }
-
-    // Call module's post init function if it exists.
-    if (init && typeof init === 'function') {
-      init();
-    }
-
-    this.api = api as API;
-  }
-
-  componentDidMount() {
-    this.mounted = true;
-  }
-
-  componentWillUnmount() {
-    this.mounted = false;
-  }
-
-  render() {
-    const { children, args } = this.props;
-    return (
-      <>
-        {React.cloneElement(children as React.ReactElement, {
-          args: {
-            ...args,
-            api: {
-              ...this.api,
-              getDocsUrl: () => 'https://storybook.js.org/docs/',
-              getUrlState: () => ({
-                queryParams: {},
-                path: '',
-                viewMode: 'story',
-                url: 'http://localhost:6006/',
-              }),
-              applyQueryParams: fn().mockName('api::applyQueryParams'),
-            },
-          },
-        })}
-      </>
-    );
-  }
-}
-
-export const MockAPIDecorator: DecoratorFunction = (Story, { args, parameters }) => (
-  <MockAPIWrapper
-    args={args}
-    initFn={initLayout}
-    initialStoryState={parameters?.initialStoryState}
-    initOptions={{ singleStory: false }}
-  >
-    <Story />
-  </MockAPIWrapper>
-);
+import { MockAPIDecorator } from './TagsFilter.story-helpers';
 
 const meta = {
   component: TagsFilter,
@@ -269,7 +144,13 @@ export const Empty: Story = {
       entries: {},
     },
   },
-  play: Clear.play,
+  play: async ({ canvas }) => {
+    const button = await canvas.findByRole('button', {}, { timeout: 3000 });
+    button.click();
+
+    const learnButton = await screen.findByText('Learn how to add tags');
+    expect(learnButton).toBeInTheDocument();
+  },
 };
 
 /** Production is equal to development now */
@@ -277,5 +158,5 @@ export const EmptyProduction: Story = {
   args: {
     ...Empty.args,
   },
-  play: Clear.play,
+  play: Empty.play,
 };
